ci: Actually run, fix lint
GitHub workflows have not been running
because the trigger was for the 'main' branch,
but the project uses a 'master' branch.

Separately,
`make lint` was broken because it expects the tools/go.mod
that specifies versions for tools dependencies.

Putting tools inside a submodule isn't absolutely necessary --
we could also put them in the root module.
I've opted for this in most of my personal and work projects
mostly because it keeps the constraints out of the main go.mod,
so the tools dependencies cannot affect constraints of your main
package.