doc: Minor cleanup
Drops references to `<script>` or ScriptBlock
since that's not needed for pikchr.

In render.go, instead of using Sprintf to generate a string,
and then writing the string to the writer,
use Fprintf to format and write directly to the writer.

Use a pikchr example instead of a mermaid example.

Depends on #2, #3